backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic[email]==2.9.2
pydantic-settings==2.6.1
slowapi==0.1.9
httpx==0.27.2
python-dotenv==1.0.1
PyJWT==2.8.0
# For production PostgreSQL (uncomment when migrating from SQLite)
# asyncpg==0.29.0
# For Redis-backed rate limiting (optional, improves multi-worker setups)
# redis==5.2.0
